refactor(platform): 提取 Unix 通用 helper 到 posix.rs 子模块
