+++ pinvou3-app/src-tauri/src/platform/os/mod.rs
@@ -5,6 +5,10 @@ pub(crate) mod macos;
 #[cfg(target_os = "windows")]
 pub(crate) mod windows;
 
+/// Unix 通用 helper（linux 与 macOS 共享）。
+#[cfg(unix)]
+pub(crate) mod posix;
+
 // `unsupported` 是「尚未实现能力」的桩集合,供 macos 借用未实现的符号(见
 // `macos/mod.rs` 的 `pub use super::unsupported::*`),并在非三大桌面平台上作为
 // 默认 platform。因此除 linux/windows 外均需声明此模块。
